switch makefile to work with new package structure
Note that since the API has changed to parse only single files
(and there are no plans to parse multiple at once),
collating the exams is now done directly in the makefile with python.

I am aware this is an extremely hacky solution, but the alternative
is to hardcode filenames into the API itself, which I definitely don't
want to do. Suggestions are welcome.

Note also that this will install gradeforge locally if it is not already
installed. I doubt this will be as controversial.