fix pytest imports
Pytest previously required the package to be installed.
There's no real reason for this and it was mostly because I wasn't sure
how to not do that. Tests have now been moved into the package proper.

See https://docs.pytest.org/en/latest/goodpractices.html#test-discovery
for more discussion on this.