select all terms by default
Previously, `download.py` would only select full-term courses by
default. This is problematic in the summer time, when most courses are
half a semester or less.

I believe '%' was recently added by the school, because it wasn't in the
original parser. I could be wrong however because I kept no notes of the
files at the time.