fix really annoying semester bug in grades
grades older than Fall 2013 were in PDF format and parsed with python;
at that time they had the semester added. since the semester was present
in the original pdf, I assumed that it was also present in the XLSX
files from later years. this was not the case.

I wanted to do this in pure python, but since there's no way to tell the
semester from the grades, the semester had to be inferred from the file
names, so it made more sense to do it in sh.

Note also that for recent semesters, it's impossible to tell what campus
a section belongs to.