add get_all_books
in order for the chrome driver to be reused, the memory reference has to
be kept; I don't know how to use it as a client-server model. as a
result, get_all_books is implemented in python, not as part of the
makefile.

on the bright side, this speeds up downloads by a factor of 10.