update sections for makefile
the sql file now depends on past semesters as well as the present
semester. parsing and cleaning had to be modified slightly to
accommodate more idiosyncrasies.

tidy is now a dependency.

additionally, recipies for sections and exams have been combined; only
dependencies differ so it didn't make sense to have them separate