make create_sql about twice as fast
the slowest operation by far is the SELECT statement to find the uid of
the section grades belong to. this operation is being reconsidered
(because sections before 2013 are not available), but until then we have
to live with it.

this skips the queries for semesters before Fall 2013, because we know
the result will always be None. additionally, we check that type errors
are because of a null result, not actual type errors.