waste time enforcing PEP 3134
https://www.python.org/dev/peps/pep-3134/ allows the syntax 'raise
Exception from e' for exception handling, to make it clear to the
developer that this is not a new exception but a refined version of the
previous one. see https://stackoverflow.com/a/24752607 for why this is
useful

the implementation is a nightmare and is very much write-once, read
never code. I considered using the `ast` module but I'm lazy.

TODO: make errors automatic, this sort of stuff should be doable with
more regular expressions.

see also http://regex.info/blog/2006-09-15/247