completely redo course parsing
closes https://github.com/jyn514/GradeForge/issues/25

this section of the parser has always been a mess because there's just
not much to go off of. There's no HTML structure at *all*, just spans
and some text at the end.

the redone parser tries to infer what a string is describing by its
content, which as you can imagine is not ideal.

additionally, there was an attribute I didn't realize I was ignoring;
some courses are only available on certain campuses