remake database on commit
previously, if you updated a source or data file, `make` would not
recreate the database. this defeats the whole point of testing.

we use `make clean && make test` instead. although it is expensive, it
prevents re-downloading files on each checkout.

the alternative of using modification dates, although cleaner, was more
technically challenging because git updates modification times on each
checkout. this does not allow caching downloads.

instead, all files are linked to the test/ directory and updated from the index.